LLDP support on Nexus 6000
Revert one of the old LLDP commits.
It tried to make 'show lldp neighbors detail' instead of 'show lldp
neighbors' on Nexus devices to determine remote interface description
and fallback to it when Port Id contained Juniper SNMP InterfaceId.

But it was a dirty hack and it worked only for some specific versions
of JunOS: modern devices transmit descriptions in Description TLVs, not
the ifName.

Also, some NX-OS versions of Nexus 6000 devices do not show 'Port ID'
field in detailed LLDP output (sic!), but show them in the standard
(brief) output.

So this commit rolls back parsing of detailed LLDP neighbor list on
Nexus devices.